backend/requirements/base.txt
# Core Django
Django==5.0.10
djangorestframework==3.14.0
django-environ==0.11.2
python-decouple==3.8

# Inertia.js
inertia-django==1.2.0

# Database
psycopg2-binary==2.9.9

# Redis & Caching
redis==5.2.0
django-redis==5.4.0

# Celery for async tasks
celery==5.4.0
django-celery-beat==2.7.0

# Image processing
Pillow==10.4.0
django-storages==1.14.4

# Payments
stripe==10.12.0

# Security & Rate Limiting
django-ratelimit==4.1.0
django-cors-headers==4.5.0

# Utilities
python-slugify==8.0.4
